refactor: s/stdlib/corelib/
