+++ src/corelib/bool.rs
@@ -1,5 +1,5 @@
 use crate::hir::*;
-//use crate::stdlib::create_method;
+//use crate::corelib::create_method;
 
 pub fn create_methods() -> Vec<SkMethod> {
     vec![
