+++ src/corelib/int.rs.erb
@@ -1,7 +1,7 @@
 // vim: set ft=rust:
 // NOTE: .rs IS AUTO-GENERATED BY .erb
 use crate::hir::*;
-use crate::stdlib::create_method;
+use crate::corelib::create_method;
 
 <%
 # Macro to define binary operator method
