+++ src/corelib/shiika_internal_memory.rs
@@ -1,6 +1,6 @@
 //use inkwell::values::*;
 use crate::hir::*;
-use crate::stdlib::create_method;
+use crate::corelib::create_method;
 
 pub fn create_class_methods() -> Vec<SkMethod> {
     vec![
