+++ src/corelib/shiika_internal_ptr.rs
@@ -1,6 +1,6 @@
 //use inkwell::values::*;
 use crate::hir::*;
-use crate::stdlib::create_method;
+use crate::corelib::create_method;
 
 pub fn create_methods() -> Vec<SkMethod> {
     vec![
