+++ src/corelib/string.rs
@@ -2,7 +2,7 @@
 use std::collections::HashMap;
 use crate::ty;
 use crate::hir::*;
-use crate::stdlib::create_method;
+use crate::corelib::create_method;
 
 pub fn create_methods() -> Vec<SkMethod> {
     vec![
