+++ src/lib.rs
@@ -4,7 +4,7 @@ pub mod ty;
 pub mod parser;
 pub mod hir;
 pub mod code_gen;
-pub mod stdlib;
+pub mod corelib;
 pub mod type_checking;
 pub mod error;
 pub mod names;
